+++ trishul_snmp/__init__.py
@@ -21,6 +21,13 @@ from trishul_snmp.notify.events import (
     decode_notification,
 )
 from trishul_snmp.notify.listener import V2cNotificationListener
+from trishul_snmp.responder.rules import (
+    CounterRule,
+    RandomNumericRule,
+    SimulationRule,
+    TimestampRule,
+    UptimeRule,
+)
 from trishul_snmp.responder.server import V2cResponder
 from trishul_snmp.responder.sources import (
     CallbackObjectSource,
@@ -57,6 +64,7 @@ __all__ = [
     "CallbackObjectSource",
     "Counter32Value",
     "Counter64Value",
+    "CounterRule",
     "EndOfMibViewValue",
     "ErrorStatus",
     "Gauge32Value",
@@ -68,25 +76,29 @@ __all__ = [
     "NoSuchInstanceValue",
     "NoSuchObjectValue",
     "NullValue",
+    "NotificationEvent",
+    "NotificationMemberBinding",
     "ObjectIdentifierValue",
     "OID",
     "OidMatch",
     "OctetStringValue",
     "OpaqueValue",
     "ProtocolError",
+    "RandomNumericRule",
     "RequestTimeoutError",
     "ResponderSource",
     "Response",
-    "NotificationEvent",
-    "NotificationMemberBinding",
+    "SimulationRule",
     "SocketAddress",
     "SnmpValue",
     "SnmpValueType",
     "TimeTicksValue",
+    "TimestampRule",
     "TranslationError",
     "TransportError",
     "UnknownOidError",
     "UnknownSymbolError",
+    "UptimeRule",
     "V2cManager",
     "V2cNotificationListener",
     "V2cNotifier",
@@ -97,4 +109,4 @@ __all__ = [
     "load_bundle",
 ]
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"